+++ app/assets/stylesheets/local/app.scss
@@ -71,6 +71,10 @@
   margin: -1rem 0 2rem;
 }
 
+.about-section {
+  margin-top: 50px;
+}
+
 // This will hide the Sentry user-feedback attribution. There is a setting
 // to hide it but currently it is not working, so this is an alternative.
 div.sentry-error-embed-wrapper p.powered-by {
