+++ app/controllers/about_controller.rb
@@ -1,0 +1,9 @@
+class AboutController < ApplicationController
+  def contact; end
+
+  def cookies; end
+
+  def privacy; end
+
+  def terms_and_conditions; end
+end
