+++ app/controllers/home_controller.rb
@@ -1,7 +1,3 @@
 class HomeController < ApplicationController
-  def contact; end
-
-  def cookies; end
-
   def miam_exemptions; end
 end
