+++ app/views/layouts/_footer_links.html.erb
@@ -2,13 +2,13 @@
   <li>
     <%= link_to t('.help'), 'https://www.gov.uk/help', class: 'ga-pageLink', data: {ga_category: 'footer', ga_label: 'help'} %>
   </li>
-  <!--<li>-->
-    <!--<%# link_to t('.contact'), contact_page_path, class: 'ga-pageLink', data: {ga_category: 'footer', ga_label: 'contact'} %>-->
-  <!--</li>-->
   <li>
-    <%= link_to t('.cookies'), cookies_page_path, class: 'ga-pageLink', data: {ga_category: 'footer', ga_label: 'cookies'} %>
+    <%= link_to t('.cookies'), about_cookies_path, class: 'ga-pageLink', data: {ga_category: 'footer', ga_label: 'cookies'} %>
   </li>
   <li>
-    <%= link_to t('.terms_and_conditions'), 'https://www.gov.uk/help/terms-conditions', class: 'ga-pageLink', data: {ga_category: 'footer', ga_label: 'terms-conditions'} %>
+    <%= link_to t('.privacy'), about_privacy_path, class: 'ga-pageLink', data: {ga_category: 'footer', ga_label: 'privacy'} %>
+  </li>
+  <li>
+    <%= link_to t('.terms_and_conditions'), about_terms_and_conditions_path, class: 'ga-pageLink', data: {ga_category: 'footer', ga_label: 'terms and conditions'} %>
   </li>
 </ul>
